Add restart functionality to SDL2 frontend
This is a bit messy. It'd make a lot of sense to actually do this
within the main game loop, since this is functionality we would want
to exist for every frontend.

This requires storing the initial values we set the game with, which
we don't currently do.

Will consider #7 still open until this functionality exists. Reading
from a file each restart is too hackish.