Add initial support for sound effects
Closes #4.

SDL_mixer was the right choice in this case. Using the default queue
audio results in sound effects buffering too much and manual callbacks
are a bit too much of a pain to use.

Currently this reads the specific sound effect data from memory. Not
interested in reading from files for the moment.

Sound effects used are taken from NullpoMino. May change in the future.