Adjust periodic draw to actually work
This allows a draw phase to occur every n game ticks instead of having
to perform a complete redraw every game tick. A draw now is defined in
terms of game ticks instead of ms to ensure we are always in sync with
out draws.

Fixed some general ini parsing problems as well to better output
unknown options.