Refactor to allow shared entry point
The main entry point is now found in `fsMain`. Some extra functions have
been added to the required interface to better support some required
features. This change makes sense on a conceptual level. Before the core
fs code was being treated more like a library by each of the frontends.
This resulted in more duplication and differences in the core gameplay
across each.

Core gameplay is now shared regardless of frontend and it should make it
much easier and more straightforward now to add a number of features
such as #15.

Closes #32.