Add font caching for glyphs
This adds font caching using the SDL_FontCache library.
 * https://github.com/grimfang4/SDL_FontCache

This nets about a 10-15% performance improvement in the main game loop.
Frame ticks take about 1ms at the moment so there was never a pressing
concern for this, but the less CPU used the better at the end of the
day.

This also changes the text colour to something less "vibrant".

Closes #37.