Switch to a 'Keys Per Tetrimino' based finesse
The previous finesse system was not really accurate at all. This one is
a KPT based for now since it is very simple to implement. I would like
to implement a fault-based system but need to work out the tables and
how to fit in rotations and the like first.

Closes #13.