Add initial replay support
This adds the base requirements for replays. A single replayed is
currently stored in the "fs.replay" file. This is fairly messy and still
needs refinements for some edge cases + extension for multiple replays
at any one time.

The actual playback is consistent however and it serves as a useful
stepping stone + tool for checking the previous game at least.

See #8.