Add CMake build system
CMake can now be used to build the project. This performs incremental
builds with proper header dependency support and is much quicker than
building normally. The old makefile is still present for users who do
not want to install CMake.

This is also a useful step in providing more cross-platform support.