Improve CMake build script + clean up project
The CMake script should now be platform-agnostic (untested). It has been
extended to create required directories on first run and a `purge`
target is added to remove all personal files.

HACKING.md was removed since it didn't explain much.

CHANGELOG.md was removed since this was a duplicate of the information
found on the releases page.

fs.ini now resides in an examples directory and the user-specified
fs.ini is not added to the vcs anymore.