Remove unnecessary header files and hide log impl
Reducing the prevalence of std headers is a good thing. Ideally we
should be able to perform a build which uses at most the following:

 * stdbool.h
 * stdlib.h
 * string.h

This is mainly for the best possible portability and ease of porting to
systems where other less common stdlib features may not be present or an
alternative is better suited.