Modify engine to be compilable freestanding
This requires a small number of custom libc implementations for when a
hosted compiler is not being used.

Currently we cannot get any randomness in a freestanding solution. This
will need to be worked with at a later time when the kernel itself has
a sufficient source of randomness and can provide this.

See #61.