Add double-buffered tty printing
This adds a new suite of functions prefixed `ttyb`. These act exactly
like the `tty` functions except that these draw to a backbuffer and not
directly to the screen. In order to show the changes, a call to
`tty_flip` must be performed.

These buffered functions update the cursor in the same manner as the
standard ones do.