Replace hiscore/replay csv flat-file with sqlite3
This is a bit easier to work with and makes bumping versions/adding new
fields a little easier. A single file is also easier to manage from a
maintainability point of view.

NOTE: terminal frontend is updated, but sdl2 frontend has yet to be. Due
to the change in how replays work, there is some pending work remaining
to determine how we want to show available replays.